requirements.txt
python-telegram-bot[job-queue]==20.3
pillow
